requirements.txt
channels==2.4.0
channels-redis==2.4.1
Django==2.2.8
django-filter==2.2.0
django-phonenumber-field==4.0.0
django-push-notifications==1.6.1
djangorestframework==3.11.0
djangorestframework-gis==0.15
GDAL==3.0.2
psycopg2==2.8.4
tinytag==1.2.2
twilio==6.35.0
django-recaptcha==2.0.6
pathlib
phonenumbers
bson
django-recurrence
python-memcached
